requirements.txt
aiofiles==22.1.0
aiomysql==0.1.1
aioredis==2.0.1
aiosmtplib==1.1.7
aiosqlite==0.17.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
Authlib==1.1.0
bcrypt==4.0.0
blinker==1.5
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
databases==0.6.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.82.0
fastapi-mail==1.1.4
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.3
